Added DICOM mask image.
Use DicomDisplayImage.SetMask( Texture2D texture ) to set a mask
which will be drawn over the current DICOM image. Alpha blending is
used, so pixels which are transparent in the mask will show the
DICOM, pixels in the mask which are NOT transparent will be drawn
over the DICOM according to their alpha value.